Move PeerRef and EndpointPeerRef to peer_ref.rs
Separate out these structs into their own file to clean
up the mod.rs file.
